libspdm: fix Yocto build failure for shared libraries
Fixes build errors when compiling shared libspdm libraries using
the BUILD_LINUX_SHARED_LIB option in Yocto for OpenSSL.

This commit fixes the following:
 1. When ENABLE_BINARY_BUILD is set, always use the existing OpenSSL
   binaries from the build environment.
 2. Conflicting data type error for time_t

+++ os_stub/openssllib/include/crt_support.h
@@ -214,7 +214,7 @@ typedef char *LIBSPDM_VA_LIST;
 
 typedef size_t u_int;
 #if defined(__GNUC__) && !defined(__MINGW64__)
-typedef size_t time_t; /* time_t is 4 bytes for 32bit machine and 8 bytes for 64bit machine */
+typedef long long time_t; /* time_t is 8 bytes for 32bit machine and 64bit machine */
 #endif
 typedef uint8_t __uint8_t;
 typedef uint8_t sa_family_t;
